requirements.txt
pyserial
ttkbootstrap
matplotlib
pandas
pyinstaller
openpyxl